Move album page to the separate folder
